Outsource header/content parsing from store.rs to util.rs for reusability
